Apply to the current implementations
